vasomim/requirements.txt


timm>=0.9.0
numpy<2.0
Pillow>=9.0
opencv-python-headless>=4.8
cupy-cuda12x>=13.0
tqdm>=4.60
matplotlib>=3.5
tensorboard>=2.14
